GlassHurricaneCandleHolder logos doubleSided
A duplicate surface has been placed inside the glass, and the logos have been applied there using the glTF properties alphaMode:Mask and doubleSided, so the logos can be seen on both the inside and outside of the glass. The rest of the glass has been divided into two materials, to preserve the bands of rough glass above and below the logos, and to prevent the logo roughness from affecting the glass over the logos.
The baseColorTexture has been edited to add an alpha channel for the logos.